requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
djangorestframework==3.9.4
gunicorn==20.0.0
Markdown==3.1.1
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.4
psycopg2-binary==2.7.7
